rename test fixture, slim down base armi test runner
